Review issues: move statement validation to FunctionProvider, marked MergeTableStatement and MergeSelectStatement as open
